DS-716: Fix menu close handling
