+++ src/components/events-discovery/components/Facets/MobileFacetFilter.jsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 import { useCurrentRefinements } from 'react-instantsearch';
 import { useFacets } from './useFacets';
 import { MobileParentFilter } from '../Filters/MobileParentFilter';
@@ -11,11 +11,27 @@ export const MobileFacetFilter = ({ attribute, onCloseMenu }) => {
 
   const count = useMemo(() => items?.[0]?.refinements?.length ?? 0, [items]);
 
+  const [isOpen, setIsOpen] = useState(false);
+
+  const open = useCallback(() => {
+    setIsOpen(true);
+  }, [setIsOpen]);
+
+  const close = useCallback(() => {
+    setIsOpen(false);
+  }, [setIsOpen]);
+
   return (
     <MobileParentFilter
+      isOpen={isOpen}
+      onOpen={open}
+      onClose={close}
       label={facet.label}
       count={count}
-      onCloseMenu={onCloseMenu}
+      onCloseMenu={() => {
+        close();
+        onCloseMenu();
+      }}
     >
       <MobileFacetList attribute={attribute} />
     </MobileParentFilter>
